Linux port: bring up libese.so + eseutil + devlibtest unit tests
Initial Linux/clang port milestone. The full LLVM toolchain is wired up
(clang/libc++/libc++abi/libunwind, statically linked into binaries) and
the OS layer is reimplemented behind the windows-shim. Engine code,
eseutil, and the seven in-scope devlibtest unit tests now build cleanly
and the unit tests pass: combined ~59 million assertions, zero failures.

Per-suite results: SYNCUNIT 253,250; RESMGRUNIT 52,106,180;
COLLECTIONUNIT 7,261,460; CcLayerUnit 101; STATUNIT 36; IterQueryUnit 6;
ERRVALIDATOR clean.

Major load-bearing changes:

- windows-shim/: new tree of Win32 API headers (winnt.h, winerror.h,
  strsafe.h, etc.) so engine source compiles unchanged. Heterogeneous
  min/max overloads (gated by _ESE_HETEROGENEOUS_MINMAX_DEFINED so the
  shim and osstd_.hxx agree on a single definition). HRESULT typedef
  is `int` not `long` (on x86_64 Linux long is 64-bit, splits
  ErrFromStrsafeHr's mangling).

- dev/ese/src/os/posix/: 22 winapi_*.cxx files plus *_posix.cxx
  per-subsystem implementations backing the windows-shim with pthread
  + a heap-allocated KObject pattern. memory_posix g_dwPageReserveGran
  matches Windows' GetSystemInfo (64K, not 4K page size). 0-byte
  aligned alloc returns a valid pointer (Win32 _aligned_malloc
  semantics). wprintf/fwprintf go through a hand-rolled UTF-16-aware
  path; libc's would print garbage under -fshort-wchar.

- dev/ese/src/sync/sync_posix.cxx: OSSYNC namespace POSIX impls
  (page memory, kernel semaphores, processor info, CLS/PLS, lifecycle).
  cbCacheLine = 32 to match the cresmgr cbRESHeader invariant.

- published/inc/sync.hxx: every _WIN64-only atomic-pointer overload
  (AtomicExchangePointer, AtomicReadPointer, AtomicCompareExchangePointer,
  AtomicRead<__int64>) extended to also fire on __LP64__. Without this
  the Linux build silently truncates pointers to 32 bits through every
  atomic — caught by SYNCUNIT and would have broken every multi-threaded
  engine path.

- published/inc/os.hxx: clang/Linux arm uses `class INST;` forward decl
  instead of `typedef unsigned char INST;` so cross-layer call sites
  (UlParam, UtilReportEvent, ErrNORMCheckLocaleName, ...) mangle the
  same on both sides under the Itanium ABI.

- published/inc/cc.hxx: alloca builtin macro, sprintf_s, _strtoui64/i64,
  INT64/UINT64 typedefs, LONGLONG=long long.

- published/inc/{logapi,flushmap,kvpstore}.hxx: in-class const-init
  static members switched to `static constexpr` so they're implicitly
  inline (without it, ODR-use forces a real symbol that lives nowhere).

- Build system:
  * libese.so wraps osposix + sync in --whole-archive and gets
    -fvisibility=default so eseutil can resolve every OS layer entry
    point through libese.so (no per-binary state duplication).
  * osposix.a compiles with -fvisibility=default and
    -fno-visibility-inlines-hidden so __forceinline functions in the
    OS layer (PefLastThrow etc.) are exported.
  * static libc++/libc++abi/libunwind via -nostdlib++ + the `_pic.a`
    variant for shared libs vs the regular `.a` for executables.
  * lld via -fuse-ld=lld (full LLVM toolchain, no GCC ld).
  * test/ subtree ungated for non-Windows; bf/, oslayer/, samples/,
    blue/, embeddedunittest, eselibwithtests, and most of resmgr/ kept
    Windows-only because they pull in Win32 surface (ETW, VirtualAlloc,
    cl-driven .def gen) that the shim doesn't cover.

+++ dev/ese/src/ese/_log/logredo.cxx
@@ -15249,7 +15249,7 @@ CheckCheckpoint:
 
             // set the initial log gen to something sane.
             m_lgenInitial = ( m_lgenInitial > pcheckpointT->checkpoint.le_lgposCheckpoint.le_lGeneration )
-                ?   pcheckpointT->checkpoint.le_lgposCheckpoint.le_lGeneration
+                ?   LONG( pcheckpointT->checkpoint.le_lgposCheckpoint.le_lGeneration )
                   : m_lgenInitial;
 
             m_pinst->m_isdlInit.FixedData().sInitData.lgposRecoveryStartMin = pcheckpointT->checkpoint.le_lgposCheckpoint;
